[GEN-638] bash history 검사관련 이슈 해결 - pod가 생성된지 일주일이 지났으며 검사 결과가 False(7일이상 사용하지않음)면 pod를 삭제 - checkCreateTime   - pod가 생성된지 7일이상 지났는지 확인하고 결과 반환 - garbagecollector.py   - manage함수에 pod삭제 조건 추

+++ garbagecollector.py
@@ -49,7 +49,7 @@ class GarbageCollector():
                 # p_obj.saveDataToCSV()
                 p_obj.saveProcessDataToDB()
 
-                if not result_history:  # 7일이상 사용하지않으면 false 반환
+                if not result_history and p_obj.checkCreateTime:  # 7일이상 사용하지않으면 false 반환
                     p_obj.insert_DeleteReason('GC_h')
                     p_obj.save_DeleteReason_to_DB()
                     self.deletePod(p_name)  # pod 삭제
